requirements.txt
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.7.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pypdfium2==4.30.0
pytesseract==0.3.13
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
